Based on an idea and patch of Gerd von Egidy (thanks!):
added extended performance data for check_multi, containing
- status data for OK, WARNING, CRITICAL, UNKNOWN
- overall state

This performance data is not added to the check_multi block,
but to a new one named check_multi_extended. This allows to 
keep the old RRDs and create a new one for the extended data.
(thanks for this nice idea to Joerg Linge)