Passive feeded service definitions can be created automatically
Service definitions for passive feeded services had to be created
with a extra call with a special report mode.
This will be simplified by a automatic mode, which checks the existence
of a service definition file at a specific directory and creates a new
service definition if needed.

Directory tree:
<nagiosdir> - etc - check_multi - feed_passive - host1 - service1.cfg
                                               |       + service2.cfg
                                               + host2 - service1.cfg
                                                       + service2.cfg