Report mode 32768 (HIDEIFOK) now hides OK states also in ASCII mode
The original idea by Timo Kempkens only covered the HTML mode, when
the condensed javascript mode is not in effect. The ASCII version was
only implemented in order to provide the same behaviour for tests.

Joel Brooks requested a different ASCII implementation: independent
from the overall state the report mode 32768 determines if OK state
child checks are displayed or not.