New timeout handling for child checks: full time range used now
In the past the timeout handling was very cautious: if a child check
came near the global timeout, it was not executed, if its child timeout
overlaps the global timeout.
This is especially a problem for long running child checks which need
long timeouts. The global timeout in this case needs to be at least
the double length of the child timeout and often collides with the
global nagios service_timeout.
The new mimic now allows to run child checks up to the parent check
timeout. If a child checks timeout overlaps the global timeout, it
is shortened accordingly.