Notification example: splitted service_output non HTML / HTML
When you specify extinfo_in_status, the normal splitting of the
check_multi output with one header line and one line per child
check is converted into HTML with a <br> instead of a newline.

Since this broke the notification example, I splitted the first line
into an ASCII part and a HTML part (behind the <br>), which now is
placed into the HTML attachment.
Thanks to Tobias Appel for the hint.