Filter mode with 'check_multi -f -' now
New versions of openssh let the perl pipe operator '-p' fail to
properly determine STDIN input. Therefore a SSH connection will
keep check_multi waiting indefinitely, until the timeout breaks it.

The error detection in this case is quite annoying, so it's better
to provide a check_multi option to clearly define, if STDIN input
is expected.

The new option is called 'check_multi -f -', where the hyphen '-'
stands for STDIN.

Thanks to Yury Pakhomenko for the error report.