Fixed macro handling to be able to deal with special characters
The new tag format also needs a new macro handling. Former macros
like $TAG$ were easy to manage, but what's about $this is a tag$?
Or more: isn't this perl snippet also a macro: 'my $a=1; $a++;'
Containing '$a=1; $'?
The fix solves this decisions on a quite practical way: if there is
a replacement for such a macro, it will be used. So if you define
export MULTI_this\ is\ a\ tag=123 or pass it via command line like
this: '-s this\ is\ a\ tag=123', you will be able to substitute
$this is a tag$ within check_multi command files.