Again removed names string in parent check data
Storing the names string was quite a bad idea. It's only a interim
data variable which contains all check command strings in order to
compute a hash out of them.
This var isn't needed for anything else, but blows up the payload
of a check, at least in XML report mode. So get rid of this crap ;)