tmp_dir with more tolerant permissions
We've returned to 41777 for the default permissions for tmp_dir:
Everybody can read and write, but only his own files.
The 750 has been experienced as too rigid especially for the
first attempts with check_multi. If more rigid permissions
are needed in certain circumstances, it can be set explicitely
via the configure option --with-tmp_dir_permissions.