Instant execution of a child check with '-i TAG'
When feeding child checks into passive services, the users might want
to reschedule such a passive service to get an instant result.
check_multi supports this with the new option '-i TAG'. The child
check with the given TAG will be executed instantaneously and the
RC of the child check will be returned as the check_multi RC.

Note the following service definition snippet for your passive service:
define service {
	...
	check_command		check_multi!-i TAG
        check_interval		0
	passive_checks_enabled	1
	active_checks_enabled	1
}