Added feed_passive perfdata with [ plugin ] at the end
When passive services are feeded with multiple child checks data,
its plugin types were not passed to PNP. Typically the plugin is
check_dummy, which prevents to select the proper template.

PNP parses a appended '[plugin]' at the end of the performance data
string for this purpose. No reason for check_multi to not add
this '[plugin]' at the end to the checkresults perfdata.