Default child check timeout increased to 11s
The default Nagios plugin timeout is per convention at 10s. The default
check_multi child check timeout was also 10s. When a plugin timed out
after 10s the check_multi timeout starts at the very same moment, and
creates a race condition, where it's not clear if the plugins RC and
output are passed correctly.

Thanks to Stan Leveau for the example and hint:
http://sourceforge.net/mailarchive/message.php?msg_id=27747352