Fixed perfdata column for livestatus command
The livestatus command now also can pass performance data to
check_multi. There is one problem remaining: livestatus does not
provide the name of the plugin which generated the perfdata.
We take the check_command string instead, so if you want to assign
a special template to the perf_data, rework your check_command name
accordingly or use the name of a plugin as check_command name.
e Please enter the commit message for your changes. Lines starting