Removed full [ STATE ] messages from status line
If a state rule is applied and it matches, only the [ STATE ] will
be shown in the first line, not the full rule. Especially for long
rules it is sufficient when the matching rules are shown in the
extended info.
With the report option -r 1 only [ STATE ] is shown, while with
the report option -r 16 the full matching rule is shown.