Introduced history for numerical values
History means that check_multi keeps track of numerical values for
e.g. the last day. All such values are stored in an ASCII file in
tmp_dir/history_<host>_<service>.txt.
The data is structured in a ring buffer. During each push old and
superfluous entries are cleaned up.

The history is controlled by the parameter history, which is per
default 0 (off). If it's 1 (on), a history is stored for each
numerical value. There are also attributes to finetune the history:
- history_maxsize   - maximum number of history values to store
- history_maxage    - maximal age of history values
- history_minvalues - minimum number of values for a valid result,
                      below a result gets state UNKNOWN

The existing data can be used for new state macros like
- AVERAGE_LAST(N)
- AVERAGE_FROM_TO(TIME_FROM,TIME_TO)
- WEIGHTED_AVERAGE_LAST(N)
- WEIGHTED_AVERAGE_FROM_TO(TIME_FROM,TIME_TO)
The non-weighted macros calculate the data points relative to the
time amount they're valid for, so that asynchronous, non regular
measuring is also computed correctly.
The weighted macros use a time shift value to give newer values a
higher precedence than older values.