Save also history values which are not currently active
Normally a history value is bound to a child check. If we take the
example of processes, a process may be running or not. And history
data of a particular process should not be deleted if it is not
running for a short time.

check_multi now saves / stores history data even if it is not active,
limited only by the maximum age and maximum number of values.